requirements.txt
fastapi
starlette[full]
jinja2
python-multipart>=0.0.9
aiofiles
uvicorn[standard]
pydantic<2
python-dotenv
pandas
numpy
pyarrow
scikit-learn
xgboost
lightgbm
torch
torchmetrics
faiss-cpu
sentence-transformers
pytesseract
Pillow
opencv-python
matplotlib
plotly
tqdm
websockets
httpx
ccxt
redis
pyyaml
sqlalchemy
alembic
joblib
optuna
